+++ assets/js/platformer3x/BossItem.js
@@ -1,6 +1,6 @@
 import GameControl from './GameControl.js';
 import GameEnv from './GameEnv.js';
-import JumpPlatform from './JumpPlatform.js';
+import JumpPlatform from './PlatformJump.js';
 
 export class BossItem extends JumpPlatform {
     constructor(canvas, image, data, xPercentage, yPercentage, name) {
