+++ assets/js/platformer3x/GameSetup.js
@@ -7,36 +7,36 @@ import BackgroundParallax from './BackgroundParallax.js';
 import BackgroundTransitions from './BackgroundTransitions.js';
 import BackgroundSnow from './BackgroundSnow.js';
 import Platform from './Platform.js';
-import JumpPlatform from './JumpPlatform.js';
+import JumpPlatform from './PlatformJump.js';
 import PlayerHills from './PlayerHills.js';
 import PlayerWinter from './PlayerWinter.js';
 import PlayerMini from './PlayerMini.js';
 import PlayerQuidditch from './PlayerQuidditch.js';
-import Goomba from './Goomba.js';
+import Goomba from './EnemyGoomba.js';
 import FlyingGoomba from './FlyingGoomba.js';
 import BlockPlatform from './BlockPlatform.js';
-import SpawnPlatform from './SpawnPlatform.js';
-import MovingPlatform from './MovingPlatform.js'
+import SpawnPlatform from './PlatformSpawn.js';
+import MovingPlatform from './PlatformMoving.js'
 import Mushroom from './Mushroom.js';
 import MagicBeam from './MagicBeam.js';
 import ChocoFrog from './ChocoFrog.js';
 import Coin from './Coin.js';
 import GameControl from './GameControl.js';
-import Owl from './Owl.js';
-import Snowman from './Snowman.js';
-import Cerberus from './Cerberus.js';
+import Owl from './FlyingOwl.js';
+import Snowman from './EnemySnowman.js';
+import Cerberus from './EnemyCerberus.js';
 import PlayerGreece from './PlayerGreece.js';
 import FinishLine from './FinishLine.js';
 import Lava from './Lava.js';
-import Dragon from './Dragon.js';
+import Dragon from './FlyingDragon.js';
 import Star from './Star.js';
-import Dementor from './Dementor.js';
-import Draco from './Draco.js';
+import Dementor from './FlyingDementor.js';
+import Draco from './EnemyDraco.js';
 import Boss from './Boss.js';
-import Jellyfish from './Jellyfish.js';
-import Penguin from './Penguin.js';
+import Jellyfish from './FlyingJellyfish.js';
+import Penguin from './EnemyPenguin.js';
 import PlayerIce from './PlayerIce.js';
-import FlyingIsland from './FlyingIsland.js';
+import FlyingIsland from './PlatformFlyingIsland.js';
 import PlayerBaseOneD from './PlayerBaseOneD.js';
 import PlayerZombie from './PlayerZombie.js';
 import BossItem from './BossItem.js';
